Remove unused API interfaces
Resolves #362
